refactor: cleanup the bloat to files
